codecheck fix

+++ services/src/form_data_mgr.cpp
@@ -1893,7 +1893,7 @@ ErrCode FormDataMgr::HandleFormRemoveObserver(const std::string hostBundleName, 
 int32_t FormDataMgr::GetTempFormsCount(int32_t &formCount)
 {
     std::lock_guard<std::recursive_mutex> lock(formTempMutex_);
-    formCount = tempForms_.size();
+    formCount = static_cast<int32_t>(tempForms_.size());
     HILOG_DEBUG("%{public}s, current exist %{public}d temp forms in system", __func__, formCount);
     return ERR_OK;
 }
